perf(web): stage watch interaction loading
