+++ apps/web/src/components/FloatingSearchController.tsx
@@ -1,0 +1,628 @@
+"use client"
+
+import {
+  useCallback,
+  useEffect,
+  useMemo,
+  useRef,
+  useState,
+  type ReactNode,
+} from "react"
+import { createPortal } from "react-dom"
+import type { Route } from "next"
+import { usePathname, useRouter } from "next/navigation"
+import { useTranslations } from "next-intl"
+
+import { runSearch } from "@/lib/search-actions"
+import { getSearchLanguageOptions } from "@/lib/search-language-actions"
+import type { SearchActionResultSource, SearchResult } from "@/lib/search"
+import {
+  MAX_SEARCH_LANGUAGE_FILTERS,
+  groupSearchLanguagesByRegion,
+  normalizeSearchLanguageEnglishNames,
+  stripLanguageFromSearchQuery,
+  type SearchLanguageCountrySuggestion,
+  type SearchLanguageOption,
+} from "@/lib/search-language"
+import { writeSearchLanguagePreferenceSlug } from "@/lib/search-language-preference-client"
+import { buildSearchUrl } from "@/lib/search-url"
+import { parseWatchPath } from "@/lib/routes"
+import {
+  FloatingSearchContext,
+  type FloatingSearchContextValue,
+} from "./FloatingSearchContext"
+
+import { SearchOverlay } from "./SearchOverlay"
+
+const SEARCH_PAGE_SIZE = 10
+
+type ActiveSearchSignature = {
+  query: string
+  languageKey: string
+  routeLanguageSlug: string | null
+  resultSource: SearchActionResultSource
+  nextOffset: number
+}
+
+export type FloatingSearchControllerProps = {
+  open: boolean
+  closing: boolean
+  query: string
+  setOpen: (open: boolean) => void
+  setQuery: (query: string) => void
+  resetToken?: number
+  children?: ReactNode
+}
+
+export function FloatingSearchController({
+  open,
+  closing,
+  query,
+  setOpen,
+  setQuery: setQueryState,
+  resetToken = 0,
+  children,
+}: FloatingSearchControllerProps) {
+  const tSearchOverlay = useTranslations("SearchOverlay")
+  const router = useRouter()
+  // usePathname() returns the app-relative path (no basePath prefix). The
+  // router.replace() call auto-prefixes basePath, so feeding it
+  // window.location.pathname (which includes basePath) would double-prefix
+  // on every search. usePathname() does NOT force the Full Route Cache
+  // deopt that useSearchParams() would.
+  const pathname = usePathname()
+
+  // Whether the modal was opened via URL hydration (vs user click). Gates a
+  // shorter skeleton threshold so the URL-hydrated blank window is less jarring.
+  const hydratedOpenRef = useRef<boolean>(false)
+
+  const [results, setResults] = useState<SearchResult[]>([])
+  const [displayResults, setDisplayResults] = useState<SearchResult[]>([])
+  const [exiting, setExiting] = useState(false)
+  const [resultsKey, setResultsKey] = useState(0)
+  const [hasMore, setHasMore] = useState(false)
+  const [loading, setLoading] = useState(false)
+  const [showSkeleton, setShowSkeleton] = useState(false)
+  const [loadingMore, setLoadingMore] = useState(false)
+  const [error, setError] = useState<string | null>(null)
+  const [searched, setSearched] = useState(false)
+  const [resultSource, setResultSource] =
+    useState<SearchActionResultSource | null>(null)
+  const [algoliaSearchEnabled, setAlgoliaSearchEnabled] = useState(false)
+  const [languageOptions, setLanguageOptions] = useState<
+    SearchLanguageOption[]
+  >([])
+  const [languageCountrySuggestion, setLanguageCountrySuggestion] =
+    useState<SearchLanguageCountrySuggestion | null>(null)
+  const [recommendedLanguage, setRecommendedLanguage] =
+    useState<SearchLanguageOption | null>(null)
+  const [languageOptionsLoading, setLanguageOptionsLoading] = useState(false)
+  const [languageOptionsError, setLanguageOptionsError] = useState<
+    string | null
+  >(null)
+  const [languageFacets, setLanguageFacets] = useState<Record<string, number>>(
+    {},
+  )
+  const [selectedLanguageEnglishNames, setSelectedLanguageEnglishNames] =
+    useState<string[]>([])
+  const [selectedLanguageRegionByName, setSelectedLanguageRegionByName] =
+    useState<Record<string, string>>({})
+
+  const requestIdRef = useRef(0)
+  const languageOptionsRequestIdRef = useRef(0)
+  const skeletonTimerRef = useRef<ReturnType<typeof setTimeout> | null>(null)
+  const loadingMoreRef = useRef(false)
+  const loadMoreRunIdRef = useRef(0)
+  const displayResultsRef = useRef<SearchResult[]>([])
+  const languageOptionsRef = useRef<SearchLanguageOption[]>([])
+  const queryRef = useRef("")
+  const selectedLanguageEnglishNamesRef = useRef<string[]>([])
+  const activeSearchSignatureRef = useRef<ActiveSearchSignature | null>(null)
+  useEffect(() => {
+    displayResultsRef.current = displayResults
+  }, [displayResults])
+  useEffect(() => {
+    languageOptionsRef.current = languageOptions
+  }, [languageOptions])
+  useEffect(() => {
+    selectedLanguageEnglishNamesRef.current = selectedLanguageEnglishNames
+  }, [selectedLanguageEnglishNames])
+
+  const setQuery = useCallback(
+    (nextQuery: string) => {
+      queryRef.current = nextQuery
+      setQueryState(nextQuery)
+    },
+    [setQueryState],
+  )
+
+  // Cancel any pending skeleton timer on unmount.
+  useEffect(() => {
+    return () => {
+      if (skeletonTimerRef.current) clearTimeout(skeletonTimerRef.current)
+    }
+  }, [])
+
+  const [portalReady, setPortalReady] = useState(false)
+  useEffect(() => {
+    setPortalReady(true)
+  }, [])
+
+  const languageGroups = useMemo(
+    () => groupSearchLanguagesByRegion(languageOptions),
+    [languageOptions],
+  )
+  const routeLanguageSlug = useMemo(() => {
+    const parsed = parseWatchPath(pathname)
+    if (
+      parsed.kind === "localized-home" ||
+      parsed.kind === "video" ||
+      parsed.kind === "episode"
+    ) {
+      return parsed.lang
+    }
+    return null
+  }, [pathname])
+
+  const refreshLanguageOptions = useCallback(
+    async (
+      availableLanguageFacets?: Record<string, number>,
+    ): Promise<SearchLanguageOption[]> => {
+      const thisRequest = ++languageOptionsRequestIdRef.current
+      setLanguageOptionsLoading(true)
+      setLanguageOptionsError(null)
+
+      try {
+        const response = await getSearchLanguageOptions({
+          availableLanguageFacets,
+        })
+        if (languageOptionsRequestIdRef.current !== thisRequest) {
+          return response.ok ? response.options : []
+        }
+        setAlgoliaSearchEnabled(response.algoliaEnabled)
+
+        if (response.ok) {
+          setLanguageOptions(response.options)
+          setLanguageCountrySuggestion(response.countrySuggestion)
+          setRecommendedLanguage(response.recommendedLanguage)
+          return response.options
+        }
+
+        setLanguageOptions([])
+        setLanguageCountrySuggestion(null)
+        setRecommendedLanguage(null)
+        setLanguageOptionsError(response.error.message)
+        return []
+      } catch {
+        if (languageOptionsRequestIdRef.current !== thisRequest) return []
+        setLanguageOptions([])
+        setLanguageCountrySuggestion(null)
+        setRecommendedLanguage(null)
+        setLanguageOptionsError("Language options are unavailable.")
+        return []
+      } finally {
+        if (languageOptionsRequestIdRef.current === thisRequest) {
+          setLanguageOptionsLoading(false)
+        }
+      }
+    },
+    [],
+  )
+
+  useEffect(() => {
+    if (!open) return
+    void refreshLanguageOptions(
+      Object.keys(languageFacets).length > 0 ? languageFacets : undefined,
+    )
+  }, [open, languageFacets, refreshLanguageOptions])
+
+  const maybeSetLanguageFacets = useCallback(
+    (
+      nextLanguageFacets: Record<string, number>,
+      activeLanguageEnglishNames: readonly string[],
+    ): void => {
+      if (activeLanguageEnglishNames.length > 0) return
+      setLanguageFacets(nextLanguageFacets)
+    },
+    [],
+  )
+
+  const search = useCallback(
+    async (
+      q: string,
+      options?: { languageEnglishNames?: string[] },
+    ): Promise<void> => {
+      const trimmed = q.trim()
+      setQuery(q)
+      const activeLanguageEnglishNames =
+        options?.languageEnglishNames ?? selectedLanguageEnglishNamesRef.current
+      const activeLanguageKey = searchLanguageKey(activeLanguageEnglishNames)
+
+      // Bump the request id immediately so any in-flight request (from a prior
+      // call in any branch — exit-animation, Apollo, or clear) fails its
+      // freshness check when it resolves.
+      const thisRequest = ++requestIdRef.current
+      loadMoreRunIdRef.current += 1
+      loadingMoreRef.current = false
+      activeSearchSignatureRef.current = null
+      setLoadingMore(false)
+
+      // URL sync — preserves any existing params (utm_*, etc.) and strips ?q=
+      // when the query is empty. Use usePathname() (app-relative) so
+      // router.replace's auto basePath prefix isn't applied twice.
+      const currentParams =
+        typeof window !== "undefined"
+          ? new URLSearchParams(window.location.search)
+          : new URLSearchParams()
+      const nextUrl = buildSearchUrl(pathname, currentParams, trimmed)
+      router.replace(nextUrl as Route)
+
+      if (!trimmed) {
+        if (displayResultsRef.current.length > 0) {
+          setExiting(true)
+          await new Promise<void>((resolve) => setTimeout(resolve, 200))
+          if (requestIdRef.current !== thisRequest) return
+          setExiting(false)
+        }
+        if (requestIdRef.current !== thisRequest) return
+        setResults([])
+        setDisplayResults([])
+        setHasMore(false)
+        setSearched(false)
+        setError(null)
+        setResultSource(null)
+        activeSearchSignatureRef.current = null
+        return
+      }
+
+      if (displayResultsRef.current.length > 0) {
+        setExiting(true)
+        await new Promise<void>((resolve) => setTimeout(resolve, 200))
+        if (requestIdRef.current !== thisRequest) return
+        setExiting(false)
+        setDisplayResults([])
+      }
+
+      setLoading(true)
+      setError(null)
+      setSearched(true)
+      if (skeletonTimerRef.current) clearTimeout(skeletonTimerRef.current)
+      const skeletonThreshold = hydratedOpenRef.current ? 150 : 500
+      skeletonTimerRef.current = setTimeout(() => {
+        if (requestIdRef.current === thisRequest) setShowSkeleton(true)
+      }, skeletonThreshold)
+
+      try {
+        const currentLanguageOptions =
+          languageOptionsRef.current.length > 0
+            ? languageOptionsRef.current
+            : await refreshLanguageOptions(
+                Object.keys(languageFacets).length > 0
+                  ? languageFacets
+                  : undefined,
+              )
+        if (requestIdRef.current !== thisRequest) return
+
+        const data = await runSearch({
+          query: trimmed.slice(0, 200),
+          limit: SEARCH_PAGE_SIZE,
+          offset: 0,
+          languageEnglishNames: activeLanguageEnglishNames,
+          languageOptions: currentLanguageOptions,
+          routeLanguageSlug,
+        })
+
+        if (requestIdRef.current !== thisRequest) return
+
+        if (!data.ok) {
+          setResults([])
+          setDisplayResults([])
+          setHasMore(false)
+          setResultSource(data.resultSource)
+          activeSearchSignatureRef.current = null
+          setError(tSearchOverlay("searchFailed"))
+          setAlgoliaSearchEnabled(data.resultSource === "algolia")
+          if (data.languageFacets) {
+            maybeSetLanguageFacets(
+              data.languageFacets,
+              activeLanguageEnglishNames,
+            )
+          }
+          return
+        }
+
+        const newResults = data.results
+        setResults(newResults)
+        setDisplayResults(newResults)
+        setResultsKey((k) => k + 1)
+        setHasMore(data.hasMore)
+        setResultSource(data.resultSource)
+        activeSearchSignatureRef.current = {
+          query: trimmed.slice(0, 200),
+          languageKey: activeLanguageKey,
+          routeLanguageSlug,
+          resultSource: data.resultSource,
+          nextOffset: data.nextOffset ?? newResults.length,
+        }
+        setAlgoliaSearchEnabled(data.resultSource === "algolia")
+        if (data.languageFacets) {
+          maybeSetLanguageFacets(
+            data.languageFacets,
+            activeLanguageEnglishNames,
+          )
+        }
+      } catch {
+        if (requestIdRef.current === thisRequest) {
+          activeSearchSignatureRef.current = null
+          setError(tSearchOverlay("searchFailed"))
+        }
+      } finally {
+        // Only clear loading state for the winning request — otherwise a
+        // stale response's finally would drop the active spinner mid-fetch.
+        if (requestIdRef.current === thisRequest) {
+          if (skeletonTimerRef.current) clearTimeout(skeletonTimerRef.current)
+          setShowSkeleton(false)
+          setLoading(false)
+        }
+      }
+    },
+    [
+      languageFacets,
+      maybeSetLanguageFacets,
+      pathname,
+      refreshLanguageOptions,
+      routeLanguageSlug,
+      router,
+      setQuery,
+      tSearchOverlay,
+    ],
+  )
+
+  const loadMore = useCallback(async (): Promise<void> => {
+    const expectedSignature = activeSearchSignatureRef.current
+    if (!expectedSignature) return
+    const currentQuery = queryRef.current.trim().slice(0, 200)
+    const currentLanguageKey = searchLanguageKey(
+      selectedLanguageEnglishNamesRef.current,
+    )
+    if (
+      expectedSignature.query !== currentQuery ||
+      expectedSignature.languageKey !== currentLanguageKey ||
+      expectedSignature.routeLanguageSlug !== routeLanguageSlug ||
+      expectedSignature.resultSource !== resultSource
+    ) {
+      return
+    }
+
+    // Synchronous guard against double-fire before React commits `disabled`.
+    if (loadingMoreRef.current) return
+    loadingMoreRef.current = true
+    const thisLoadMoreRun = ++loadMoreRunIdRef.current
+    setLoadingMore(true)
+    setError(null)
+    // Capture the current search's request id; bail out of the append if a
+    // new search supersedes us mid-fetch.
+    const thisRequest = requestIdRef.current
+    try {
+      const data = await runSearch({
+        query: currentQuery,
+        limit: SEARCH_PAGE_SIZE,
+        offset: expectedSignature.nextOffset,
+        languageEnglishNames: selectedLanguageEnglishNamesRef.current,
+        languageOptions: languageOptionsRef.current,
+        routeLanguageSlug,
+      })
+      if (requestIdRef.current !== thisRequest) return
+      if (data.resultSource !== expectedSignature.resultSource) {
+        setError(tSearchOverlay("loadMoreFailed"))
+        return
+      }
+      if (!data.ok) {
+        setError(tSearchOverlay("loadMoreFailed"))
+        setAlgoliaSearchEnabled(data.resultSource === "algolia")
+        return
+      }
+      setResults((prev) => [...prev, ...data.results])
+      setDisplayResults((prev) => [...prev, ...data.results])
+      setHasMore(data.hasMore)
+      setResultSource(data.resultSource)
+      activeSearchSignatureRef.current = {
+        ...expectedSignature,
+        nextOffset:
+          data.nextOffset ?? expectedSignature.nextOffset + data.results.length,
+      }
+      setAlgoliaSearchEnabled(data.resultSource === "algolia")
+    } catch {
+      if (requestIdRef.current === thisRequest) {
+        setError(tSearchOverlay("loadMoreFailed"))
+      }
+    } finally {
+      if (loadMoreRunIdRef.current === thisLoadMoreRun) {
+        loadingMoreRef.current = false
+        setLoadingMore(false)
+      }
+    }
+  }, [resultSource, routeLanguageSlug, tSearchOverlay])
+
+  const toggleSearchLanguage = useCallback(
+    (option: SearchLanguageOption, regionName?: string): void => {
+      const selected = selectedLanguageEnglishNamesRef.current.some(
+        (language) => language === option.englishName,
+      )
+      if (
+        !selected &&
+        selectedLanguageEnglishNamesRef.current.length >=
+          MAX_SEARCH_LANGUAGE_FILTERS
+      ) {
+        return
+      }
+      const nextLanguages = selected
+        ? selectedLanguageEnglishNamesRef.current.filter(
+            (language) => language !== option.englishName,
+          )
+        : normalizeSearchLanguageEnglishNames([
+            ...selectedLanguageEnglishNamesRef.current,
+            option.englishName,
+          ])
+
+      selectedLanguageEnglishNamesRef.current = nextLanguages
+      setSelectedLanguageEnglishNames(nextLanguages)
+      setSelectedLanguageRegionByName((prev) => {
+        const next = { ...prev }
+        if (selected) {
+          delete next[option.englishName]
+        } else if (regionName) {
+          next[option.englishName] = regionName
+        }
+        return next
+      })
+
+      if (!selected && option.publicSlug) {
+        writeSearchLanguagePreferenceSlug(option.publicSlug)
+      }
+
+      const nextQuery = selected
+        ? query
+        : stripLanguageFromSearchQuery(option.englishName, query)
+      void search(nextQuery, { languageEnglishNames: nextLanguages })
+    },
+    [query, search],
+  )
+
+  const selectSearchLanguage = useCallback(
+    (option: SearchLanguageOption, regionName?: string): void => {
+      const nextLanguages = [option.englishName]
+      selectedLanguageEnglishNamesRef.current = nextLanguages
+      setSelectedLanguageEnglishNames(nextLanguages)
+      setSelectedLanguageRegionByName(
+        regionName ? { [option.englishName]: regionName } : {},
+      )
+
+      if (option.publicSlug) {
+        writeSearchLanguagePreferenceSlug(option.publicSlug)
+      }
+    },
+    [],
+  )
+
+  const clearSearchLanguages = useCallback((): void => {
+    selectedLanguageEnglishNamesRef.current = []
+    setSelectedLanguageEnglishNames([])
+    setSelectedLanguageRegionByName({})
+    if (query.trim().length > 0) {
+      void search(query, { languageEnglishNames: [] })
+    }
+  }, [query, search])
+
+  const closeAndKeepQuery = useCallback(() => {
+    setOpen(false)
+  }, [setOpen])
+
+  const resetTokenRef = useRef(resetToken)
+  useEffect(() => {
+    if (resetToken === 0 || resetTokenRef.current === resetToken) return
+    resetTokenRef.current = resetToken
+    void search("")
+  }, [resetToken, search])
+
+  // On first client mount, seed state from ?q= in the URL.
+  const didHydrateRef = useRef(false)
+  useEffect(() => {
+    if (didHydrateRef.current) return
+    didHydrateRef.current = true
+    if (typeof window === "undefined") return
+    const seeded = new URLSearchParams(window.location.search).get("q") ?? ""
+    const trimmed = seeded.slice(0, 200)
+    if (trimmed.length === 0) return
+    hydratedOpenRef.current = true
+    setQuery(trimmed)
+    setOpen(true)
+    void search(trimmed)
+  }, [search, setOpen, setQuery])
+
+  const value = useMemo<FloatingSearchContextValue>(
+    () => ({
+      open,
+      closing,
+      query,
+      results,
+      displayResults,
+      exiting,
+      resultsKey,
+      hasMore,
+      loading,
+      showSkeleton,
+      loadingMore,
+      error,
+      searched,
+      resultSource,
+      algoliaSearchEnabled,
+      languageOptions,
+      languageGroups,
+      languageCountrySuggestion,
+      recommendedLanguage,
+      languageOptionsLoading,
+      languageOptionsError,
+      selectedLanguageEnglishNames,
+      selectedLanguageRegionByName,
+      setOpen,
+      setQuery,
+      search,
+      loadMore,
+      toggleSearchLanguage,
+      selectSearchLanguage,
+      clearSearchLanguages,
+      closeAndKeepQuery,
+    }),
+    [
+      open,
+      closing,
+      query,
+      results,
+      displayResults,
+      exiting,
+      resultsKey,
+      hasMore,
+      loading,
+      showSkeleton,
+      loadingMore,
+      error,
+      searched,
+      resultSource,
+      algoliaSearchEnabled,
+      languageOptions,
+      languageGroups,
+      languageCountrySuggestion,
+      recommendedLanguage,
+      languageOptionsLoading,
+      languageOptionsError,
+      selectedLanguageEnglishNames,
+      selectedLanguageRegionByName,
+      setOpen,
+      setQuery,
+      search,
+      loadMore,
+      toggleSearchLanguage,
+      selectSearchLanguage,
+      clearSearchLanguages,
+      closeAndKeepQuery,
+    ],
+  )
+
+  const modalChromeHidden = open || closing
+
+  return (
+    <FloatingSearchContext.Provider value={value}>
+      {children}
+      {portalReady && modalChromeHidden
+        ? createPortal(<SearchOverlay />, document.body)
+        : null}
+    </FloatingSearchContext.Provider>
+  )
+}
+
+function searchLanguageKey(languageEnglishNames: readonly string[]): string {
+  return normalizeSearchLanguageEnglishNames(languageEnglishNames).join(
+    "\u0001",
+  )
+}
