+++ apps/web/src/components/watch/WatchPageClient.tsx
@@ -49,7 +49,12 @@ import {
   writeSubtitlePreference,
 } from "@/lib/subtitle-preference-client"
 import { resolvePosterUrl } from "@/lib/url"
-import { loadWatchLanguageOptions } from "@/lib/watch-language-actions"
+import {
+  getCachedWatchLanguageOptions,
+  loadWatchInteraction,
+  loadWatchLanguageOptionsForVideo,
+  scheduleWatchInteractionWarmup,
+} from "@/lib/watch-interaction-loader"
 
 function resolveSubtitleSlug(
   preferred: string | null,
@@ -218,6 +223,11 @@ export function WatchPageClient({
   const [downloadError, setDownloadError] = useState<string | null>(null)
   const downloadPendingRef = useRef(false)
   const videoSlug = video.slug ?? ""
+  const [enabledModalChunks, setEnabledModalChunks] = useState({
+    download: false,
+    language: false,
+    share: false,
+  })
   const [languageOptionsState, setLanguageOptionsState] =
     useState<LanguageOptionsState>({
       status: "idle",
@@ -227,7 +237,16 @@ export function WatchPageClient({
 
   useEffect(() => {
     languageOptionsPendingRef.current = false
-    setLanguageOptionsState({ status: "idle", variants: [] })
+    const cached = videoSlug ? getCachedWatchLanguageOptions(videoSlug) : null
+    setLanguageOptionsState(
+      cached
+        ? { status: "ready", variants: cached }
+        : { status: "idle", variants: [] },
+    )
+  }, [videoSlug])
+
+  useEffect(() => {
+    return scheduleWatchInteractionWarmup({ videoSlug })
   }, [videoSlug])
 
   const loadLanguageOptions = useCallback(async () => {
@@ -235,10 +254,16 @@ export function WatchPageClient({
     if (languageOptionsPendingRef.current) return
     if (languageOptionsState.status === "ready") return
 
+    const cached = getCachedWatchLanguageOptions(videoSlug)
+    if (cached) {
+      setLanguageOptionsState({ status: "ready", variants: cached })
+      return
+    }
+
     languageOptionsPendingRef.current = true
     setLanguageOptionsState({ status: "loading", variants: [] })
     try {
-      const variants = await loadWatchLanguageOptions({ videoSlug })
+      const variants = await loadWatchLanguageOptionsForVideo(videoSlug)
       setLanguageOptionsState({ status: "ready", variants })
     } catch {
       setLanguageOptionsState({ status: "error", variants: [] })
@@ -249,6 +274,8 @@ export function WatchPageClient({
 
   const openDownload = useCallback(async () => {
     if (downloadPendingRef.current) return
+    setEnabledModalChunks((prev) => ({ ...prev, download: true }))
+    void loadWatchInteraction("download").catch(() => {})
     downloadPendingRef.current = true
     setDownloadPending(true)
 
@@ -270,10 +297,16 @@ export function WatchPageClient({
     }
   }, [tDownloadButton])
   const openLanguage = useCallback(() => {
+    setEnabledModalChunks((prev) => ({ ...prev, language: true }))
+    void loadWatchInteraction("language").catch(() => {})
     setModalState("language")
     void loadLanguageOptions()
   }, [loadLanguageOptions])
-  const openShare = useCallback(() => setModalState("share"), [])
+  const openShare = useCallback(() => {
+    setEnabledModalChunks((prev) => ({ ...prev, share: true }))
+    void loadWatchInteraction("share").catch(() => {})
+    setModalState("share")
+  }, [])
   const closeModal = useCallback(() => setModalState("none"), [])
 
   // Pause the video whenever any modal (search / language / download / share)
@@ -342,42 +375,48 @@ export function WatchPageClient({
         initialTranscript={initialTranscript}
       />
 
-      <DownloadModal
-        open={modalState === "download"}
-        downloads={downloadsForModal}
-        videoTitle={video.title ?? null}
-        posterUrl={posterUrl}
-        durationSeconds={variant.duration ?? null}
-        languageName={variant.language?.name ?? null}
-        variantId={variant.documentId}
-        videoSlug={videoSlug}
-        onClose={closeModal}
-      />
-      <LanguagePickerModal
-        open={modalState === "language"}
-        variants={languageOptionsState.variants}
-        currentLanguageSlug={currentLanguageSlug}
-        videoSlug={videoSlug}
-        playerRef={playerRef}
-        onClose={closeModal}
-        subtitles={subtitles}
-        currentSubtitleEnabled={subtitleEnabled}
-        currentSubtitleSlug={subtitleSlug}
-        onSubtitleChange={handleSubtitleChange}
-        languageOptionsLoading={languageOptionsState.status === "loading"}
-        languageOptionsError={languageOptionsState.status === "error"}
-        onRetryLanguageOptions={loadLanguageOptions}
-      />
-      <ShareModal
-        open={modalState === "share"}
-        videoSlug={videoSlug}
-        currentLanguageSlug={currentLanguageSlug}
-        videoTitle={video.title ?? null}
-        videoDescription={video.snippet ?? video.description ?? null}
-        posterUrl={posterUrl}
-        playbackId={variant.muxVideo?.playbackId ?? null}
-        onClose={closeModal}
-      />
+      {enabledModalChunks.download ? (
+        <DownloadModal
+          open={modalState === "download"}
+          downloads={downloadsForModal}
+          videoTitle={video.title ?? null}
+          posterUrl={posterUrl}
+          durationSeconds={variant.duration ?? null}
+          languageName={variant.language?.name ?? null}
+          variantId={variant.documentId}
+          videoSlug={videoSlug}
+          onClose={closeModal}
+        />
+      ) : null}
+      {enabledModalChunks.language ? (
+        <LanguagePickerModal
+          open={modalState === "language"}
+          variants={languageOptionsState.variants}
+          currentLanguageSlug={currentLanguageSlug}
+          videoSlug={videoSlug}
+          playerRef={playerRef}
+          onClose={closeModal}
+          subtitles={subtitles}
+          currentSubtitleEnabled={subtitleEnabled}
+          currentSubtitleSlug={subtitleSlug}
+          onSubtitleChange={handleSubtitleChange}
+          languageOptionsLoading={languageOptionsState.status === "loading"}
+          languageOptionsError={languageOptionsState.status === "error"}
+          onRetryLanguageOptions={loadLanguageOptions}
+        />
+      ) : null}
+      {enabledModalChunks.share ? (
+        <ShareModal
+          open={modalState === "share"}
+          videoSlug={videoSlug}
+          currentLanguageSlug={currentLanguageSlug}
+          videoTitle={video.title ?? null}
+          videoDescription={video.snippet ?? video.description ?? null}
+          posterUrl={posterUrl}
+          playbackId={variant.muxVideo?.playbackId ?? null}
+          onClose={closeModal}
+        />
+      ) : null}
       {questionPanelEnabled ? (
         <WatchQuestionPanel
           enabled={questionPanelEnabled}
